Changed order of nouns and properties.
Now objects are 0-7, nouns are 8-15, and the rest are
properties and verbs.  This clears the way to add more
properties and verbs.

I may move "IS" to 8 since there doesn't seem to be a way
to put anything meaningful in that tile.  even if I put
scenery graphics there, it would not fit in the background
bits of PF%().