fixed bug where a trail of YOU could disappear
If you had three YOU objects in a row, and moved left or up,
all of them would disappear except the one at the head.  Moving
right or down worked correctly.  So I changed it to traverse
the U% array in the opposite direction when DX is negative.
I could have used a FOR-NEXT loop but then Basic-BOSS would
not have been able to compile it, and it still would have been
difficult to understand.

The loop also only works if DX never changes inside the movement
subroutine.