improved space compression more
originally I avoided runs of 2 or 3 spaces to avoid putting a
double-quote in the DATA statement, but instead I put in a scheme
where it replaces the double-quote with a pi character.

The program data increase probably won't make up for the compression,
but remember the goal is to make the DATA statement fit in 80
characters.

Trailing spaces can be trimmed manually.